Fix types

+++ presentation-compiler/test/dotty/tools/pc/utils/JRE.scala
@@ -7,7 +7,7 @@ object JRE:
 
     javaVersion match
       case Some(version) if version.startsWith("1.8") => 8
-      case Some(version) => version
+      case Some(version) => version.toInt // it is better to crash during tests than to run incorrect suite
       case None => 8
